backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.3
python-dotenv==1.0.1
google-cloud-firestore==2.16.0
openai==1.35.10
httpx==0.27.0
